+++ include/common/monolish_view1D.hpp
@@ -68,6 +68,9 @@ public:
   size_t size() const { return range; }
   size_t get_nnz() const { return range; }
 
+  size_t get_first() const { return first; }
+  size_t get_last() const { return last; }
+  size_t get_offset() const { return first; }
   void set_first(size_t i) { first = i; }
 
   void set_last(size_t i) {
@@ -78,7 +81,7 @@ public:
   size_t get_device_mem_stat() const { return target.get_device_mem_stat(); }
 
   Float *data() { return target_data; }
-  Float *data() const { return data(); }
+  Float *data() const { return target_data; }
 
   void print_all(bool force_cpu = false) const;