+++ src/blas/matrix/matvec/crs_matvec.cpp
@@ -5,8 +5,8 @@ namespace monolish {
 
 namespace {
 // double ///////////////////
-template <typename VEC>
-void Dmatvec_core(const matrix::CRS<double> &A, const VEC &x, VEC &y) {
+template <typename VEC1, typename VEC2>
+void Dmatvec_core(const matrix::CRS<double> &A, const VEC1 &x, VEC2 &y) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -23,6 +23,8 @@ void Dmatvec_core(const matrix::CRS<double> &A, const VEC &x, VEC &y) {
   const int n = A.get_col();
   const int nnz = A.get_nnz();
   double *yd = y.data();
+  const size_t xoffset = x.get_offset();
+  const size_t yoffset = y.get_offset();
 
   if (A.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU // gpu
@@ -44,8 +46,8 @@ void Dmatvec_core(const matrix::CRS<double> &A, const VEC &x, VEC &y) {
 #pragma omp target data use_device_ptr(xd, yd, vald, rowd, cold)
     {
       internal::check_CUDA(cusparseDcsrmv(sp_handle, trans, m, n, nnz, &alpha,
-                                          descr, vald, rowd, cold, xd, &beta,
-                                          yd));
+                                          descr, vald, rowd, cold, xd + xoffset,
+                                          &beta, yd + yoffset));
     }
 #else
     throw std::runtime_error("error USE_GPU is false, but gpu_status == true");
@@ -55,8 +57,8 @@ void Dmatvec_core(const matrix::CRS<double> &A, const VEC &x, VEC &y) {
 #if MONOLISH_USE_MKL
     const double alpha = 1.0;
     const double beta = 0.0;
-    mkl_dcsrmv("N", &m, &n, &alpha, "G__C", vald, cold, rowd, rowd + 1, xd,
-               &beta, yd);
+    mkl_dcsrmv("N", &m, &n, &alpha, "G__C", vald, cold, rowd, rowd + 1,
+               xd + xoffset, &beta, yd + yoffset);
 
     // OSS
 #else
@@ -64,9 +66,9 @@ void Dmatvec_core(const matrix::CRS<double> &A, const VEC &x, VEC &y) {
     for (int i = 0; i < (int)A.get_row(); i++) {
       double ytmp = 0.0;
       for (int j = A.row_ptr[i]; j < A.row_ptr[i + 1]; j++) {
-        ytmp += vald[j] * xd[A.col_ind[j]];
+        ytmp += vald[j] * (xd + xoffset)[A.col_ind[j]];
       }
-      y[i] = ytmp;
+      y[i + yoffset] = ytmp;
     }
 #endif
   }
@@ -75,8 +77,8 @@ void Dmatvec_core(const matrix::CRS<double> &A, const VEC &x, VEC &y) {
 }
 
 // float ///////////////////
-template <typename VEC>
-void Smatvec_core(const matrix::CRS<float> &A, const VEC &x, VEC &y) {
+template <typename VEC1, typename VEC2>
+void Smatvec_core(const matrix::CRS<float> &A, const VEC1 &x, VEC2 &y) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -93,6 +95,8 @@ void Smatvec_core(const matrix::CRS<float> &A, const VEC &x, VEC &y) {
   const int n = A.get_col();
   const int nnz = A.get_nnz();
   float *yd = y.data();
+  const size_t xoffset = x.get_offset();
+  const size_t yoffset = y.get_offset();
 
   if (A.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU // gpu
@@ -114,8 +118,8 @@ void Smatvec_core(const matrix::CRS<float> &A, const VEC &x, VEC &y) {
 #pragma omp target data use_device_ptr(xd, yd, vald, rowd, cold)
     {
       internal::check_CUDA(cusparseScsrmv(sp_handle, trans, m, n, nnz, &alpha,
-                                          descr, vald, rowd, cold, xd, &beta,
-                                          yd));
+                                          descr, vald, rowd, cold, xd + xoffset,
+                                          &beta, yd + yoffset));
     }
 #else
     throw std::runtime_error("error USE_GPU is false, but gpu_status == true");
@@ -125,8 +129,8 @@ void Smatvec_core(const matrix::CRS<float> &A, const VEC &x, VEC &y) {
 #if MONOLISH_USE_MKL
     const float alpha = 1.0;
     const float beta = 0.0;
-    mkl_scsrmv("N", &m, &n, &alpha, "G__C", vald, cold, rowd, rowd + 1, xd,
-               &beta, yd);
+    mkl_scsrmv("N", &m, &n, &alpha, "G__C", vald, cold, rowd, rowd + 1,
+               xd + xoffset, &beta, yd + yoffset);
 
     // OSS
 #else
@@ -134,9 +138,9 @@ void Smatvec_core(const matrix::CRS<float> &A, const VEC &x, VEC &y) {
     for (int i = 0; i < (int)A.get_row(); i++) {
       float ytmp = 0.0;
       for (int j = A.row_ptr[i]; j < A.row_ptr[i + 1]; j++) {
-        ytmp += vald[j] * xd[A.col_ind[j]];
+        ytmp += vald[j] * (xd + xoffset)[A.col_ind[j]];
       }
-      y[i] = ytmp;
+      y[i + yoffset] = ytmp;
     }
 #endif
   }
@@ -146,8 +150,22 @@ void Smatvec_core(const matrix::CRS<float> &A, const VEC &x, VEC &y) {
 } // namespace
 
 namespace blas {
+
 void matvec(const matrix::CRS<double> &A, const vector<double> &x,
             vector<double> &y) {
+  Dmatvec_core(A, x, y);
+}
+void matvec(const matrix::CRS<double> &A, const vector<double> &x,
+            view1D<vector<double>, double> &y) {
+  Dmatvec_core(A, x, y);
+}
+void matvec(const matrix::CRS<double> &A,
+            const view1D<vector<double>, double> &x, vector<double> &y) {
+  Dmatvec_core(A, x, y);
+}
+void matvec(const matrix::CRS<double> &A,
+            const view1D<vector<double>, double> &x,
+            view1D<vector<double>, double> &y) {
   Dmatvec_core(A, x, y);
 }
 
@@ -155,6 +173,18 @@ void matvec(const matrix::CRS<float> &A, const vector<float> &x,
             vector<float> &y) {
   Smatvec_core(A, x, y);
 }
-} // namespace blas
+void matvec(const matrix::CRS<float> &A, const vector<float> &x,
+            view1D<vector<float>, float> &y) {
+  Smatvec_core(A, x, y);
+}
+void matvec(const matrix::CRS<float> &A, const view1D<vector<float>, float> &x,
+            vector<float> &y) {
+  Smatvec_core(A, x, y);
+}
+void matvec(const matrix::CRS<float> &A, const view1D<vector<float>, float> &x,
+            view1D<vector<float>, float> &y) {
+  Smatvec_core(A, x, y);
+}
 
+} // namespace blas
 } // namespace monolish