+++ src/blas/matrix/matvec/linearoperator_matvec.cpp
@@ -4,8 +4,8 @@
 namespace monolish {
 
 namespace {
-template <typename T, typename VEC>
-void matvec_core(const matrix::LinearOperator<T> &A, const VEC &x, VEC &y) {
+template <typename T, typename VEC1, typename VEC2>
+void matvec_core(const matrix::LinearOperator<T> &A, const VEC1 &x, VEC2 &y) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -20,8 +20,8 @@ void matvec_core(const matrix::LinearOperator<T> &A, const VEC &x, VEC &y) {
   logger.func_out();
 }
 
-template <typename T, typename VEC>
-void rmatvec_core(const matrix::LinearOperator<T> &A, const VEC &x, VEC &y) {
+template <typename T, typename VEC1, typename VEC2>
+void rmatvec_core(const matrix::LinearOperator<T> &A, const VEC1 &x, VEC2 &y) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -38,6 +38,7 @@ void rmatvec_core(const matrix::LinearOperator<T> &A, const VEC &x, VEC &y) {
 } // namespace
 
 namespace blas {
+
 void matvec(const matrix::LinearOperator<double> &A, const vector<double> &x,
             vector<double> &y) {
   matvec_core(A, x, y);
@@ -46,6 +47,7 @@ void rmatvec(const matrix::LinearOperator<double> &A, const vector<double> &x,
              vector<double> &y) {
   rmatvec_core(A, x, y);
 }
+
 void matvec(const matrix::LinearOperator<float> &A, const vector<float> &x,
             vector<float> &y) {
   matvec_core(A, x, y);
@@ -54,6 +56,6 @@ void rmatvec(const matrix::LinearOperator<float> &A, const vector<float> &x,
              vector<float> &y) {
   rmatvec_core(A, x, y);
 }
-} // namespace blas
 
+} // namespace blas
 } // namespace monolish