+++ src/blas/vector/axpyz.cpp
@@ -17,12 +17,15 @@ void Daxpyz_core(const F1 alpha, const F2 &x, const F3 &y, F4 &z) {
   const double *yd = y.data();
   double *zd = z.data();
   size_t size = x.size();
+  const size_t xoffset = x.get_offset();
+  const size_t yoffset = y.get_offset();
+  const size_t zoffset = z.get_offset();
 
   if (x.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for
     for (size_t i = 0; i < size; i++) {
-      zd[i] = alpha * xd[i] + yd[i];
+      zd[i + zoffset] = alpha * xd[i + xoffset] + yd[i + yoffset];
     }
 #else
     throw std::runtime_error(
@@ -31,7 +34,7 @@ void Daxpyz_core(const F1 alpha, const F2 &x, const F3 &y, F4 &z) {
   } else {
 #pragma omp parallel for
     for (size_t i = 0; i < size; i++) {
-      zd[i] = alpha * xd[i] + yd[i];
+      zd[i + zoffset] = alpha * xd[i + xoffset] + yd[i + yoffset];
     }
   }
   logger.func_out();
@@ -50,12 +53,15 @@ void Saxpyz_core(const F1 alpha, const F2 &x, const F3 &y, F4 &z) {
   const float *yd = y.data();
   float *zd = z.data();
   size_t size = x.size();
+  const size_t xoffset = x.get_offset();
+  const size_t yoffset = y.get_offset();
+  const size_t zoffset = z.get_offset();
 
   if (x.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for
     for (size_t i = 0; i < size; i++) {
-      zd[i] = alpha * xd[i] + yd[i];
+      zd[i + zoffset] = alpha * xd[i + xoffset] + yd[i + yoffset];
     }
 #else
     throw std::runtime_error(
@@ -64,7 +70,7 @@ void Saxpyz_core(const F1 alpha, const F2 &x, const F3 &y, F4 &z) {
   } else {
 #pragma omp parallel for
     for (size_t i = 0; i < size; i++) {
-      zd[i] = alpha * xd[i] + yd[i];
+      zd[i + zoffset] = alpha * xd[i + xoffset] + yd[i + yoffset];
     }
   }
   logger.func_out();
@@ -77,9 +83,69 @@ void axpyz(const double alpha, const vector<double> &x, const vector<double> &y,
            vector<double> &z) {
   Daxpyz_core(alpha, x, y, z);
 }
+void axpyz(const double alpha, const vector<double> &x, const vector<double> &y,
+           view1D<vector<double>, double> &z) {
+  Daxpyz_core(alpha, x, y, z);
+}
+void axpyz(const double alpha, const vector<double> &x,
+           const view1D<vector<double>, double> &y, vector<double> &z) {
+  Daxpyz_core(alpha, x, y, z);
+}
+void axpyz(const double alpha, const vector<double> &x,
+           const view1D<vector<double>, double> &y,
+           view1D<vector<double>, double> &z) {
+  Daxpyz_core(alpha, x, y, z);
+}
+void axpyz(const double alpha, const view1D<vector<double>, double> &x,
+           const vector<double> &y, vector<double> &z) {
+  Daxpyz_core(alpha, x, y, z);
+}
+void axpyz(const double alpha, const view1D<vector<double>, double> &x,
+           const vector<double> &y, view1D<vector<double>, double> &z) {
+  Daxpyz_core(alpha, x, y, z);
+}
+void axpyz(const double alpha, const view1D<vector<double>, double> &x,
+           const view1D<vector<double>, double> &y, vector<double> &z) {
+  Daxpyz_core(alpha, x, y, z);
+}
+void axpyz(const double alpha, const view1D<vector<double>, double> &x,
+           const view1D<vector<double>, double> &y,
+           view1D<vector<double>, double> &z) {
+  Daxpyz_core(alpha, x, y, z);
+}
 
 void axpyz(const float alpha, const vector<float> &x, const vector<float> &y,
            vector<float> &z) {
+  Saxpyz_core(alpha, x, y, z);
+}
+void axpyz(const float alpha, const vector<float> &x, const vector<float> &y,
+           view1D<vector<float>, float> &z) {
+  Saxpyz_core(alpha, x, y, z);
+}
+void axpyz(const float alpha, const vector<float> &x,
+           const view1D<vector<float>, float> &y, vector<float> &z) {
+  Saxpyz_core(alpha, x, y, z);
+}
+void axpyz(const float alpha, const vector<float> &x,
+           const view1D<vector<float>, float> &y,
+           view1D<vector<float>, float> &z) {
+  Saxpyz_core(alpha, x, y, z);
+}
+void axpyz(const float alpha, const view1D<vector<float>, float> &x,
+           const vector<float> &y, vector<float> &z) {
+  Saxpyz_core(alpha, x, y, z);
+}
+void axpyz(const float alpha, const view1D<vector<float>, float> &x,
+           const vector<float> &y, view1D<vector<float>, float> &z) {
+  Saxpyz_core(alpha, x, y, z);
+}
+void axpyz(const float alpha, const view1D<vector<float>, float> &x,
+           const view1D<vector<float>, float> &y, vector<float> &z) {
+  Saxpyz_core(alpha, x, y, z);
+}
+void axpyz(const float alpha, const view1D<vector<float>, float> &x,
+           const view1D<vector<float>, float> &y,
+           view1D<vector<float>, float> &z) {
   Saxpyz_core(alpha, x, y, z);
 }