+++ src/blas/vector/nrm1.cpp
@@ -11,12 +11,13 @@ template <typename F1> double Dnrm1_core(const F1 &x) {
   double ans = 0;
   const double *xd = x.data();
   size_t size = x.size();
+  const size_t xoffset = x.get_offset();
 
   if (x.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for reduction(+ : ans) map (tofrom: ans)
     for (size_t i = 0; i < size; i++) {
-      ans += std::abs(xd[i]);
+      ans += std::abs(xd[i + xoffset]);
     }
 #else
     throw std::runtime_error(
@@ -25,7 +26,7 @@ template <typename F1> double Dnrm1_core(const F1 &x) {
   } else {
 #pragma omp parallel for reduction(+ : ans)
     for (size_t i = 0; i < size; i++) {
-      ans += std::abs(xd[i]);
+      ans += std::abs(xd[i + xoffset]);
     }
   }
 
@@ -40,12 +41,13 @@ template <typename F1> float Snrm1_core(const F1 &x) {
   float ans = 0;
   const float *xd = x.data();
   size_t size = x.size();
+  const size_t xoffset = x.get_offset();
 
   if (x.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for reduction(+ : ans) map (tofrom: ans)
     for (size_t i = 0; i < size; i++) {
-      ans += std::abs(xd[i]);
+      ans += std::abs(xd[i + xoffset]);
     }
 #else
     throw std::runtime_error(
@@ -54,7 +56,7 @@ template <typename F1> float Snrm1_core(const F1 &x) {
   } else {
 #pragma omp parallel for reduction(+ : ans)
     for (size_t i = 0; i < size; i++) {
-      ans += std::abs(xd[i]);
+      ans += std::abs(xd[i + xoffset]);
     }
   }
 
@@ -65,11 +67,18 @@ template <typename F1> float Snrm1_core(const F1 &x) {
 } // namespace
 
 namespace blas {
-double nrm1(const vector<double> &x) { return Dnrm1_core(x); }
-void nrm1(const vector<double> &x, double &ans) { ans = nrm1(x); }
 
+double nrm1(const vector<double> &x) { return Dnrm1_core(x); }
+double nrm1(const view1D<vector<double>, double> &x) { return Dnrm1_core(x); }
 float nrm1(const vector<float> &x) { return Snrm1_core(x); }
+float nrm1(const view1D<vector<float>, float> &x) { return Snrm1_core(x); }
+
+void nrm1(const vector<double> &x, double &ans) { ans = nrm1(x); }
+void nrm1(const view1D<vector<double>, double> &x, double &ans) {
+  ans = nrm1(x);
+}
 void nrm1(const vector<float> &x, float &ans) { ans = nrm1(x); }
+void nrm1(const view1D<vector<float>, float> &x, float &ans) { ans = nrm1(x); }
 
 } // namespace blas