+++ src/blas/vector/sum.cpp
@@ -11,12 +11,13 @@ template <typename F1> double Dsum_core(const F1 &x) {
   double ans = 0;
   const double *xd = x.data();
   size_t size = x.size();
+  const size_t xoffset = x.get_offset();
 
   if (x.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for reduction(+ : ans) map (tofrom: ans)
     for (size_t i = 0; i < size; i++) {
-      ans += xd[i];
+      ans += xd[i + xoffset];
     }
 #else
     throw std::runtime_error(
@@ -25,7 +26,7 @@ template <typename F1> double Dsum_core(const F1 &x) {
   } else {
 #pragma omp parallel for reduction(+ : ans)
     for (size_t i = 0; i < size; i++) {
-      ans += xd[i];
+      ans += xd[i + xoffset];
     }
   }
 
@@ -40,12 +41,13 @@ template <typename F1> float Ssum_core(const F1 &x) {
   float ans = 0;
   const float *xd = x.data();
   size_t size = x.size();
+  const size_t xoffset = x.get_offset();
 
   if (x.get_device_mem_stat() == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for reduction(+ : ans) map (tofrom: ans)
     for (size_t i = 0; i < size; i++) {
-      ans += xd[i];
+      ans += xd[i + xoffset];
     }
 #else
     throw std::runtime_error(
@@ -54,7 +56,7 @@ template <typename F1> float Ssum_core(const F1 &x) {
   } else {
 #pragma omp parallel for reduction(+ : ans)
     for (size_t i = 0; i < size; i++) {
-      ans += xd[i];
+      ans += xd[i + xoffset];
     }
   }
 
@@ -65,11 +67,16 @@ template <typename F1> float Ssum_core(const F1 &x) {
 } // namespace
 
 namespace blas {
+
 double sum(const vector<double> &x) { return Dsum_core(x); }
+double sum(const view1D<vector<double>, double> &x) { return Dsum_core(x); }
 void sum(const vector<double> &x, double &ans) { ans = sum(x); }
+void sum(const view1D<vector<double>, double> &x, double &ans) { ans = sum(x); }
 
 float sum(const vector<float> &x) { return Ssum_core(x); }
+float sum(const view1D<vector<float>, float> &x) { return Ssum_core(x); }
 void sum(const vector<float> &x, float &ans) { ans = sum(x); }
+void sum(const view1D<vector<float>, float> &x, float &ans) { ans = sum(x); }
 
 } // namespace blas